+++ src/main/java/RenovatioMod/renovatio/item/ModItems.java
@@ -3,6 +3,8 @@ package RenovatioMod.renovatio.item;
 import RenovatioMod.renovatio.item.custom.equipment.ModMeleeWeapons;
 import RenovatioMod.renovatio.item.custom.equipment.ModPlateArmor;
 import RenovatioMod.renovatio.item.custom.equipment.ModRangedWeapons;
+import RenovatioMod.renovatio.item.custom.trinket.underarmor.ModUnderArmor;
+import RenovatioMod.renovatio.item.custom.trinket.underarmor.underarmor_c.UnderArmorChest;
 
 public class ModItems {
 
@@ -12,6 +14,8 @@ public class ModItems {
         ModRangedWeapons.registerItems();
         ModMaterialItems.registerItems();
         ModPlateArmor.registerItems();
+        ModUnderArmor.registerItems();
+
 
 
     }